requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
celery==3.1.23
click==6.6
cssselect==0.9.2
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.11.1
Flask-SQLAlchemy==2.1
idna==2.1
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
kombu==3.0.35
lxml==3.6.4
MarkupSafe==0.23
mysqlclient==1.3.7
newspaper3k==0.1.7
nltk==3.2.1
Pillow==3.3.1
pkg-resources==0.0.0
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
requests-file==1.4
six==1.10.0
SQLAlchemy==1.0.15
tldextract==2.0.1
uWSGI==2.0.13.1
Werkzeug==0.11.11
